move SkeletonView to API folder
